refactor(core): move canvas parsing into the iiifParsing module
`getCanvasChoices` and `getCanvasBehaviors` read IIIF canvas JSON and
know the v2/v3 spelling difference, but lived in `components/
viewerControls.ts` -- a components module otherwise concerned with
layout: nav layout, paged grouping, visible entries, choice groups.

They now sit alongside the rest of the parsing surface in
`utils/iiifParsing.ts`, next to `getPaintingBody`, `isChoiceBody` and
`getChoiceAlternatives`, which is what they are built from. Nothing is
re-exported from the old location: a compatibility re-export would leave
two import paths for one function, which is the opposite of centralizing.

`isSinglePageCanvas` stays in `viewerControls.ts`. It asks a layout
question ("does this canvas want its own spread?") and merely consults
the parsed behaviors to answer it.

The behavioral golden caught a mistake in this move. `viewer.baseline
.test.ts` imported `getCanvasChoices` from the old path, and the broad
and deep tiers both moved before that import was fixed -- a mechanical
relocation is exactly the change that should move nothing, so the golden
did the job it was frozen for.

`api-reports/dts-any-allowlist.txt` keys by declaration file, so both
entries are repointed rather than regenerated; the count is unchanged at
79.

This is a partial consolidation, deliberately. Roughly 80 version-spelling
reads still live in `structures.ts`, `collections.ts`, `languageMap.ts`,
`getThumbnailSrc.ts`, `resolveCanvasImage.ts`, `iiifIds.ts` and
`metadataNormalization.ts`. Those parse things that are not canvases, or
resolve values rather than walking documents, and folding them in would
build one large module out of genuinely different concerns.
